dom/imptests/html/html/dom/documents/dta/test_document.embeds-document.plugins-01.html

<!DOCTYPE html>
<title>document.embeds and document.plugins</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#dom-document-embeds">
<link rel="help" href="http://www.whatwg.org/html5/#dom-document-plugins">
<link rel="stylesheet" href="/resources/testharness.css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_equals(document.embeds, document.plugins);
  assert_equals(document.embeds.length, 0);
  assert_equals(document.plugins.length, 0);
});
test(function() {
  document.body.appendChild(document.createElement("embed"));
  assert_equals(document.embeds, document.plugins);
  assert_equals(document.embeds.length, 1);
  assert_equals(document.plugins.length, 1);
  assert_equals(document.embeds[0], document.getElementsByTagName("embed")[0]);
  assert_equals(document.plugins[0], document.getElementsByTagName("embed")[0]);
});
</script>
